fico/requirements.txt
numpy>=1.24
pandas>=2.0
plotly>=5.22
ipywidgets>=8.0
scikit-learn>=1.3
nbformat>=5.10
jupyterlab>=4.0

# ---- GPU text generation + embeddings (optional but recommended) ----
# NOTE: For A100/CUDA, install the CUDA-enabled PyTorch wheel that matches your system.
# See: https://pytorch.org/get-started/locally/
torch>=2.1
transformers>=4.44
accelerate>=0.33
sentence-transformers>=3.0
tqdm>=4.66
jinja2>=3.1.0

# Module C: real RAG pipeline + NIM calls + eval
requests>=2.31
faiss-cpu>=1.8
ragas>=0.1.16
datasets>=2.20
langchain>=0.2
langchain-community>=0.2




